moved to using unsigned ints for all places where it was appropriate. basic work on polygon, passing all tests

+++ source/circle.cpp
@@ -7,10 +7,11 @@ using std::string;
 #include <utility>
 using std::pair;
 
-Circle::Circle(int radius, BoundBoxType bound_box, PointType current_point)
+Circle::Circle(unsigned int radius, BoundBoxType bound_box,
+               PointType current_point)
     : radius(radius), Shape(bound_box, current_point) {}
 
-void Circle::rotate(size_t rotation_angle) {}
+void Circle::rotate(unsigned int rotation_angle) {}
 
 void Circle::scale(double x_scale, double y_scale) {}
